@@ -1,9 +1,12 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import App from './App';
+import {cleanup, fireEvent, render, screen} from '@testing-library/react';
 
-it('renders without crashing', () => {
-  const div = document.createElement('div');
-  ReactDOM.render(<App />, div);
-  ReactDOM.unmountComponentAtNode(div);
+afterEach(cleanup);
+
+describe('App Component', () => {
+    it('renders without crashing', () => {
+        const {unmount} = render(<App/>);
+        unmount();
+    });
 });
